writing the logic of the CustomerService and writing tests for it

+++ src/StormyCommerce.Core/Entities/Catalog/Product/Brand.cs
@@ -2,7 +2,7 @@
 {
     public class Brand : BaseEntity
     {
-        public Brand(int id)
+        public Brand(long id)
         {
             Id = id;
         }
